mangrove-classifier/requirements.txt
tensorflow
torch
torchvision
numpy
matplotlib
scikit-learn
pillow
jupyter
flask
flask-cors
requests
